Add TestSalaryForHours

+++ src/test/java/edu/sabana/poob/CircleTest.java
@@ -24,7 +24,7 @@ public class CircleTest {
     public void shouldCalculateArea() {
         assertEquals(314, (int) c1.getArea());
         assertEquals(660, (int) c2.getArea());
-        assertEquals(1735, (int) Math.round(c3.getArea()));
+        assertEquals(1735, Math.round(c3.getArea()));
         assertEquals(3, (int) c4.getArea());
     }
